+++ BACK-END/Models/Consumption.cs
@@ -6,9 +6,9 @@ namespace BACK_END.Models
     {
         public int Id { get; set; }
         [Column(TypeName = "decimal(18,2)")]
-        public decimal Water { get; set; }
+        public double Water { get; set; }
         [Column(TypeName = "decimal(18,2)")]
-        public decimal Electric { get; set; }
+        public double Electric { get; set; }
         public DateTime Time { get; set; }
         public int RoomId { get; set; }
         public Room? Room { get; set; }
